:wrench: Add versions fixture
