+++ tests/files/test_update_files.py
@@ -27,20 +27,6 @@ mutation (
 """
 
 
-@pytest.fixture
-def versions(db, clients, mocker):
-    client = clients.get("Administrators")
-    study = StudyFactory()
-    file = FileFactory(study=study)
-    version = file.versions.latest("created_at")
-    version.key = open(f"tests/data/manifest.txt")
-
-    mock_resp = mocker.patch("creator.files.views._resolve_version")
-    mock_resp.return_value = (file, version)
-
-    return study, file, version
-
-
 def test_unauthed_file_mutation_query(db, clients, versions):
     """
     File mutations are not allowed without authentication
